+++ services/src/main/resources/META-INF/services/org.keycloak.broker.provider.IdentityProviderFactory
@@ -20,4 +20,4 @@ org.keycloak.broker.oidc.KeycloakOIDCIdentityProviderFactory
 org.keycloak.broker.saml.SAMLIdentityProviderFactory
 org.keycloak.broker.oauth.OAuth2IdentityProviderFactory
 org.keycloak.broker.spiffe.SpiffeIdentityProviderFactory
-org.keycloak.broker.kube.KubeIdentityProviderFactory
+org.keycloak.broker.kubernetes.KubernetesIdentityProviderFactory
